Add some new cards and enemy types.
I really need to work out some way of keeping these values in a sortable
and measurable table. Might be worth considering doing all of these in
something like excel and writing an import/export function.

For now I'm going to try to work on calculating some relative "value"
terms for both enemies and cards so I have some sense of what their
relative value to the player is.