Resize battle scene
I don't know that all that added space is helpful. It makes the scope of
the game a lot larger and is one more thing the player has to do.
Keeping all the action on a single screen is probably a better idea and
is easier to balance and build for.